feat(system): feat(spawn): fix(seedgo): audit cleanup — CLAUDE.md + README date + test_quality output_capture (DPLAN-0141 Phase 3)
