+++ src/aipass/ai_mail/README.md
@@ -5,7 +5,7 @@
 **Purpose:** Inter-agent messaging for AIPass. File-based email system that lets agents send, receive, and process messages using `@branch` addresses. No SMTP, no external services — just JSON files and symbolic routing.
 **Module:** `aipass.ai_mail`
 **Created:** 2025-11-08
-**Last Updated:** 2026-04-07
+**Last Updated:** 2026-04-22
 
 ---
